requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
beautifulsoup4==4.12.3
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
Django==5.0.3
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap4==24.1
django-crispy-forms==2.1
django-tempus-dominus==5.1.2.17
pydantic==2.6.4
pydantic_core==2.16.3
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
